Change header actions to be icons
Ref #3353, this changes the mobile app header actions to be icons across
the board. (for now it's only the info icon, but we can add more icons
as we go along) Right now, it does so on both Android and iOS but I'm
happy to add text-based top right buttons back in case they feel more
native on iOS. (I use Android so no clue)

/cc @brianlovin @uberbryn
