+++ mobile/views/TabBar/BaseStack.js
@@ -2,6 +2,7 @@
 // The basic view stack that's used on all of our screens
 // Any view that's added here can be visited from any of our tabs
 import React from 'react';
+import { withMappedNavigationProps } from 'react-navigation-props-mapper';
 import Thread from '../Thread';
 import ThreadDetail from '../ThreadDetail';
 import Community from '../Community';
@@ -10,12 +11,11 @@ import Channel from '../Channel';
 import ChannelDetail from '../ChannelDetail';
 import UserDetail from '../UserDetail';
 import User from '../User';
-import { withMappedNavigationProps } from 'react-navigation-props-mapper';
-import type { NavigationScreenConfigProps } from 'react-navigation';
-import NavigateToThreadDetails from './headerActions/NavigateToThreadDetails';
-import NavigateToCommunityDetails from './headerActions/NavigateToCommunityDetails';
-import NavigateToChannelDetails from './headerActions/NavigateToChannelDetails';
-import NavigateToUserDetails from './headerActions/NavigateToUserDetails';
+import Info from './headerActions/Info';
+import type {
+  NavigationScreenConfigProps,
+  NavigationProp,
+} from 'react-navigation';
 
 const BaseStack = {
   ThreadDetail: {
@@ -28,7 +28,17 @@ const BaseStack = {
     screen: withMappedNavigationProps(Thread),
     navigationOptions: ({ navigation }: NavigationScreenConfigProps) => ({
       headerTitle: navigation.getParam('title', null),
-      headerRight: <NavigateToThreadDetails navigation={navigation} />,
+      headerRight: (
+        <Info
+          onPress={() =>
+            navigation.navigate({
+              routeName: `ThreadDetail`,
+              key: `thread-detail-${navigation.state.key}`,
+              params: { id: navigation.state.params.id },
+            })
+          }
+        />
+      ),
     }),
   },
   CommunityDetail: {
@@ -41,7 +51,17 @@ const BaseStack = {
     screen: withMappedNavigationProps(Community),
     navigationOptions: ({ navigation }: NavigationScreenConfigProps) => ({
       headerTitle: navigation.getParam('title', null),
-      headerRight: <NavigateToCommunityDetails navigation={navigation} />,
+      headerRight: (
+        <Info
+          onPress={() =>
+            navigation.navigate({
+              routeName: `CommunityDetail`,
+              key: `community-detail-${navigation.state.key}`,
+              params: { id: navigation.state.params.id },
+            })
+          }
+        />
+      ),
     }),
   },
   ChannelDetail: {
@@ -54,14 +74,34 @@ const BaseStack = {
     screen: withMappedNavigationProps(Channel),
     navigationOptions: ({ navigation }: NavigationScreenConfigProps) => ({
       headerTitle: navigation.getParam('title', null),
-      headerRight: <NavigateToChannelDetails navigation={navigation} />,
+      headerRight: (
+        <Info
+          onPress={() =>
+            navigation.navigate({
+              routeName: `ChannelDetail`,
+              key: `channel-detail-${navigation.state.key}`,
+              params: { id: navigation.state.params.id },
+            })
+          }
+        />
+      ),
     }),
   },
   User: {
     screen: withMappedNavigationProps(User),
     navigationOptions: ({ navigation }: NavigationScreenConfigProps) => ({
       headerTitle: navigation.getParam('title', null),
-      headerRight: <NavigateToUserDetails navigation={navigation} />,
+      headerRight: (
+        <Info
+          onPress={() =>
+            navigation.navigate({
+              routeName: `UserDetail`,
+              key: `user-detail-${navigation.state.key}`,
+              params: { id: navigation.state.params.id },
+            })
+          }
+        />
+      ),
     }),
   },
   UserDetail: {
