--- mobile/views/TabBar/headerActions/NavigateToUserDetails.js
@@ -1,23 +1,0 @@
-// @flow
-import * as React from 'react';
-import type { NavigationProps } from 'react-navigation';
-import { Button } from 'react-native';
-
-type Props = {
-  navigation: NavigationProps,
-};
-
-export default class NavigateToThreadDetails extends React.Component<Props> {
-  navigate = () => {
-    const { navigation } = this.props;
-    return navigation.navigate({
-      routeName: `UserDetail`,
-      key: `user-detail-${navigation.state.key}`,
-      params: { id: navigation.state.params.id },
-    });
-  };
-
-  render() {
-    return <Button onPress={this.navigate} title="Details" />;
-  }
-}
